# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WOSP API Documentation",
        "description": "WOSP API in Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "WOSP API Server"
        }
    ],
    "paths": {
        "/stations": {
            "get": {
                "tags": [
                    "Stations"
                ],
                "summary": "Get list of Sttions",
                "description": "Returns list of Stations",
                "operationId": "getStationsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Stations"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/stations/{id}/busy": {
            "post": {
                "tags": [
                    "Stations"
                ],
                "summary": "Set Station status to busy",
                "description": "Set Station status to busy",
                "operationId": "busyStation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Station id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Station set to busy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Station"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/stations/{id}/unknown": {
            "post": {
                "tags": [
                    "Stations"
                ],
                "summary": "Set Station status to unknown",
                "description": "Set Station status to unknown",
                "operationId": "unknownStation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Station id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Station set to busy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Station"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/stations/status": {
            "get": {
                "tags": [
                    "Stations"
                ],
                "summary": "Get list of Sttions statuses",
                "description": "Returns list of Stations statuses",
                "operationId": "getStationsStatusesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "data",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "STATUS_READY"
                                                    },
                                                    "value": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/charityBoxes": {
            "get": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Get list of charity boxes",
                "description": "Returns list of charity boxes",
                "operationId": "getCharityBoxList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBoxResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Create charity box for volunteer (collector)",
                "description": "Returns information about realse charity box for volunteer (collector)",
                "operationId": "createCharityBox",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CollectorCharityBoxRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/charityBoxes/unverified": {
            "get": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Get list of unverified charity boxes",
                "description": "Returns list of charity boxes",
                "operationId": "getUnverifiedList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBoxResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/charityBoxes/verified": {
            "get": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Get list of verified charity boxes",
                "description": "Returns list of charity boxes",
                "operationId": "getVerifiedList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBoxResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/charityBoxes/verified/{id}": {
            "post": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Verify Charity Box",
                "description": "Return message of operation result",
                "operationId": "postVerifyCharityBoxById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Charity box id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/charityBoxes/unverified/{id}": {
            "post": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Unverify Charity Box",
                "description": "Return message of operation result",
                "operationId": "postUnverifyCharityBoxById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Charity box id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/charityBoxes/{id}/startCounting": {
            "post": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Start counting Charity Box",
                "description": "Return charity box instance",
                "operationId": "stratCountngCharityBoxById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Charity box id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/charityBoxes/{id}/finishCounting": {
            "post": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Finish counting Charity Box",
                "description": "Return charity box instance",
                "operationId": "finishCountngCharityBoxById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Charity box id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/collectors": {
            "get": {
                "tags": [
                    "Collectors"
                ],
                "summary": "Get list of Collectors",
                "description": "Returns list of Collectors",
                "operationId": "getCollectorList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collectors"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Collectors"
                ],
                "summary": "Insert a new Collector",
                "description": "Insert a new Collector",
                "operationId": "postCollector",
                "requestBody": {
                    "description": "Collector to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CollectorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collector created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collector"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Collector exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Istnieje już wolontariusz o podanym numerze w systemie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/collectors/{id}": {
            "get": {
                "tags": [
                    "Collectors"
                ],
                "summary": "Get collector information",
                "description": "Returns collector data",
                "operationId": "getCollectorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collector id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collector"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/collectors/{collectorIdentifier}/box/create": {
            "post": {
                "tags": [
                    "Collectors"
                ],
                "summary": "Create Charity Box for Collector",
                "description": "Create Charity Box for Collector",
                "operationId": "giveCollectorBox",
                "parameters": [
                    {
                        "name": "collectorIdentifier",
                        "in": "path",
                        "description": "Collector (volunteer) identifier id (number)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/logs": {
            "get": {
                "tags": [
                    "BoxEvent"
                ],
                "summary": "Get list of box events - logs",
                "description": "Returns list of Logs",
                "operationId": "getBoxEventsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BoxEvents"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Stations": {
                "title": "Stations",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Station"
                        }
                    }
                },
                "type": "object"
            },
...
            "Station": {
                "title": "Station",
                "description": "Station model - ",
                "properties": {
                    "station": {
                        "title": "Station",
                        "description": "Station number",
                        "type": "integer",
                        "example": "1"
                    },
                    "status": {
                        "title": "Stationary status",
                        "description": "Stationary status- active: 1, inactive: 0",
                        "type": "bit",
                        "example": "1"
                    },
                    "time": {
                        "title": "Stationary time",
                        "description": "Stationary status- active: 1, inactive: 0",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-06 12:38:42"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
...
}